Updated desktop environment
